requirements-dev.txt
pytest>=7.2.2
pytest-cov>=4.0.0
requests_mock>=1.10.0
prometheus_client>=0.18.0
python-decouple>=3.8
schedule>=1.2.1

black==23.3.0
flake8==6.0.0
isort==5.12.0
flake8-pyproject==1.2.3
